Rework to split code and isolate for page types
