+++ app/views/partials/_properties.html.erb
@@ -1,5 +1,8 @@
 <script>
   var app = {
+    page: {
+      type:  "<%= controller_name %>"
+    },
     text: {
       dialogs: {
         button_cancel: "<%= t('dialogs.button_cancel') %>",
